cuda graph embedding and output layer
