+++ megatron/core/models/hybrid/hybrid_block.py
@@ -25,7 +25,7 @@ from megatron.core.process_groups_config import ProcessGroupCollection
 from megatron.core.transformer import TransformerConfig
 from megatron.core.transformer.enums import CudaGraphScope
 from megatron.core.transformer.identity_op import IdentityOp
-from megatron.core.transformer.module import GraphableMegatronModule, MegatronModule
+from megatron.core.transformer.module import MegatronModule
 from megatron.core.transformer.spec_utils import ModuleSpec, build_module
 from megatron.core.transformer.transformer_layer import TransformerLayer
 from megatron.core.transformer.utils import sharded_state_dict_default
@@ -47,7 +47,7 @@ class HybridStackSubmodules:
     mtp_block_spec: Optional[ModuleSpec] = None
 
 
-class HybridStack(GraphableMegatronModule, MegatronModule):
+class HybridStack(MegatronModule):
     """
     Constructor for the HybridStack class.
 
@@ -205,39 +205,6 @@ class HybridStack(GraphableMegatronModule, MegatronModule):
             if layer_type == LayerSymbols.MAMBA:
                 return layer.mamba_state_shapes_per_request()
         return None
-
-    def _should_call_local_cudagraph(self, *args, **kwargs):
-        """
-        Check if we should call the local cudagraph path.
-        """
-        if (
-            not self.training
-            and hasattr(self, 'cudagraph_manager')
-            and kwargs['attention_mask'] is None
-            and (
-                kwargs.get('inference_context') is not None
-                or kwargs.get('inference_params') is not None
-            )
-            and CudaGraphScope.full_iteration_inference in self.config.cuda_graph_scope
-        ):
-            if kwargs['inference_context'].is_static_batching():
-                using_cuda_graph = kwargs['inference_context'].is_decode_only()
-            else:
-                using_cuda_graph = kwargs['inference_context'].using_cuda_graph_this_step()
-
-            if using_cuda_graph:
-                return True
-        return False
-
-    def __call__(self, *args, **kwargs):
-        if self._should_call_local_cudagraph(*args, **kwargs):
-            kwargs['hidden_states'] = (
-                kwargs['hidden_states'].unwrap()
-                if isinstance(kwargs['hidden_states'], WrappedTensor)
-                else kwargs['hidden_states']
-            )
-            return super().__call__(*args, **kwargs)[0]
-        return super().__call__(*args, **kwargs)
 
     def forward(
         self,
