+++ megatron/core/models/hybrid/hybrid_model.py
@@ -32,11 +32,13 @@ from megatron.core.utils import (
     is_using_quantization_scales,
     log_single_rank,
 )
+from megatron.core.transformer.module import GraphableMegatronModule
+from megatron.core.transformer.enums import CudaGraphScope
 
 logger = logging.getLogger(__name__)
 
 
-class HybridModel(LanguageModule):
+class HybridModel(LanguageModule, GraphableMegatronModule):
     """Hybrid language model.
 
     Args:
@@ -319,6 +321,38 @@ class HybridModel(LanguageModule):
                 for param in self.output_layer.parameters():
                     off_interface.mark_not_offloadable(param)
             self.disable_param_offloading = False
+
+    def _should_call_local_cudagraph(self, *args, **kwargs):
+        """
+        Check if we should call the local cudagraph path.
+        """
+        if (
+            not self.training
+            and hasattr(self, 'cudagraph_manager')
+            and (
+                kwargs.get('inference_context') is not None
+                or kwargs.get('inference_params') is not None
+            )
+            and CudaGraphScope.full_iteration_inference in self.config.cuda_graph_scope
+        ):
+            if kwargs['inference_context'].is_static_batching():
+                using_cuda_graph = kwargs['inference_context'].is_decode_only()
+            else:
+                using_cuda_graph = kwargs['inference_context'].using_cuda_graph_this_step()
+
+            if using_cuda_graph:
+                return True
+        return False
+
+    def __call__(self, *args, **kwargs):
+        if self._should_call_local_cudagraph(*args, **kwargs):
+            return super().__call__(*args, **kwargs)[0]
+        return super().__call__(*args, **kwargs)
+    
+    def create_mcore_cudagraph_manager(self, config):
+        if CudaGraphScope.full_iteration_inference in config.cuda_graph_scope:
+            from megatron.core.transformer.cuda_graphs import CudaGraphManager
+            self.cudagraph_manager = CudaGraphManager(config)
 
     def forward(
         self,
